idastar

+++ srcs/Board.hpp
@@ -15,7 +15,7 @@ struct Board
     Board move(int i0, int j0, int new_i, int new_j) const;
     int linear_conflicts(bool partition = false) const;
     float partition_cost() const;
-    float estimate_cost() const;
+    int estimate_cost() const;
     int get_empty_coords() const;
     bool istarget() const;
     // bool operator==(Board &rhs) const;
